Small changes based on review
